feat(verify): refactor to modular architecture with verified header support
Architecture changes:
- Split verify.ts (1000+ lines) into modular structure:
  - scripts/verify/types.ts - Shared interfaces
  - scripts/verify/cache.ts - Caching utilities
  - scripts/verify/docker.ts - Docker utilities
  - scripts/verify/runner.ts - JS execution
  - scripts/verify/parser.ts - Function file parsing
  - scripts/verify/languages/index.ts - Language registry
  - scripts/verify/languages/php.ts - PHP handler
  - scripts/verify/languages/python.ts - Python handler
- Main verify.ts is now ~250 lines CLI entry point

New features:
- Add `verified: X.Y` header support in function files
  - Functions with header are verified in CI mode (default)
  - Functions without header are skipped unless --all is used
- Add --summary flag to show verification status counts
- Add --all flag to include unverified functions

CI integration:
- Default mode only runs functions with verified: header
- Exit code 1 only for verified function failures
- Unverified functions shown as "skipped" count

Demo verified headers added:
- src/php/strings/trim.js: verified: 8.3
- src/python/math/factorial.js: verified: 3.12
